+++ frontend/src/ProtectedRoute.jsx
@@ -5,7 +5,7 @@ import { Navigate } from "react-router-dom";
 const ProtectedRoute = ({ children }) => {
   const token = useSelector((state) => state.auth.token);
 
-  return token ? children : <Navigate to="login" />;
+  return token ? children : <Navigate to="/login" />;
 };
 
 ProtectedRoute.propTypes = {
