+++ frontend/src/components/Error.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Error = styled.p`
+  color: red;
+`;
+
+export default Error;
