--- frontend/src/components/Form/Error.jsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const Error = styled.span`
-  color: var(--warning);
-  position: absolute;
-  top: 0;
-  right: 0;
-  z-index: 1;
-  font-size: var(--fs_small);
-`;
-
-export default Error;
