--- frontend/src/pages/About.jsx
@@ -1,2 +1,0 @@
-const About = () => <h1>About Us</h1>;
-export default About;
