+++ frontend/src/pages/Home.jsx
@@ -2,7 +2,6 @@ import styled from "styled-components";
 
 import Hero from "../components/Hero";
 import Features from "../components/Features";
-import homeBanner from "../assets/bank-tree.jpeg";
 
 const Main = styled.main`
   flex: 1;
@@ -11,7 +10,7 @@ const Main = styled.main`
 function Home() {
   return (
     <Main>
-      <Hero image={homeBanner} />
+      <Hero />
       <Features />
     </Main>
   );
